restructure web component
