+++ client/src/components/web/NFTs/CreateNFTs/DeployConfirmModal.jsx
@@ -2,9 +2,9 @@ import React, { useState, useRef } from 'react';
 import { useDispatch } from 'react-redux';
 import { Modal, Button, Form, FormControl } from 'react-bootstrap';
 import { Formik } from 'formik';
-import { putDeploy } from '../../../actions/deployActions';
-import { updateNFTLocalStorage } from '../../../actions/NFTActions';
-import { nftContractDeploy } from '../../../services/nftServices';
+import { putDeploy } from '../../../../actions/deployActions';
+import { updateNFTLocalStorage } from '../../../../actions/NFTActions';
+import { nftContractDeploy } from '../../../../services/nftServices';
 
 export const DeployConfirmModal = ({ show, handleClose, publicKey }) => {
 	const dispatch = useDispatch();
