+++ client/src/components/web/NFTs/CreateNFTs/NFTMintForm.jsx
@@ -3,15 +3,15 @@ import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { Button, Form, FormControl } from 'react-bootstrap';
 import { Formik, Field } from 'formik';
-import { putDeploy } from '../../../actions/deployActions';
-import { storeFile, deleteFile } from '../../../actions/fileActions';
-import { updateNFTLocalStorage } from '../../../actions/NFTActions';
-import { getSingedMintDeploy } from '../../../services/nftServices';
-import { ImagePreview } from '../../Common/Image/ImagePreview';
-import { NFT_GATEWAY } from '../../../constants/key';
-import { validateNFTMintForm } from '../../../helpers/validator';
-import { massageNFTMintFormValues } from '../../../helpers/nft';
-import { MAX_METADATA_ATTRIBUTES } from '../../../constants/nft';
+import { putDeploy } from '../../../../actions/deployActions';
+import { storeFile, deleteFile } from '../../../../actions/fileActions';
+import { updateNFTLocalStorage } from '../../../../actions/NFTActions';
+import { getSingedMintDeploy } from '../../../../services/nftServices';
+import { ImagePreview } from '../../../Common/Image/ImagePreview';
+import { NFT_GATEWAY } from '../../../../constants/key';
+import { validateNFTMintForm } from '../../../../helpers/validator';
+import { massageNFTMintFormValues } from '../../../../helpers/nft';
+import { MAX_METADATA_ATTRIBUTES } from '../../../../constants/nft';
 import { NFTModal } from '../NFTModal';
 import SelectField from './SelectField';
 import { NFTAttributeRow } from './NFTAttributeRow';
