+++ client/src/components/web/Stake/index.test.jsx
