+++ client/src/components/web/Wallets/index.test.jsx
